{{cookiecutter.project_name}}/requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
decorator==4.4.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.38.1
h11==0.8.1
httptools==0.0.13
idna==2.8
importlib-metadata==0.22
ipykernel==4.10.1
ipython==5.8.0
ipython-genutils==0.2.0
jupyter-client==5.3.2
jupyter-core==4.5.0
more-itertools==7.2.0
numpy==1.17.2
packaging==19.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
py==1.8.0
pydantic==0.32.2
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-multipart==0.0.5
pyzmq==18.1.0
simplegeneric==0.8.1
six==1.12.0
starlette==0.12.8
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
traitlets==4.3.2
uvicorn==0.9.0
uvloop==0.13.0
wcwidth==0.1.7
websockets==8.0.2
zipp==0.6.0
